+++ src/prometheus-net.Contrib/EventListeners/Adapters/PrometheusAspNetCoreCounterAdapter.cs
@@ -1,10 +1,8 @@
-﻿using System.Collections.Generic;
-using Prometheus.Contrib.Core;
-using Prometheus.Contrib.EventListeners.Counters;
+﻿using Prometheus.Contrib.EventListeners.Counters;
 
 namespace Prometheus.Contrib.EventListeners.Adapters
 {
-    internal class PrometheusAspNetCoreCounterAdapter : ICounterAdapter
+    internal class PrometheusAspNetCoreCounterAdapter : BaseAdapter
     {
         public const string EventSourceName = "Microsoft.AspNetCore.Hosting";
             
@@ -12,25 +10,5 @@ namespace Prometheus.Contrib.EventListeners.Adapters
         internal readonly MeanCounter TotalRequests = new MeanCounter("total-requests","aspnetcore_requests_total", "Total Requests");
         internal readonly MeanCounter CurrentRequests = new MeanCounter("current-requests", "aspnetcore_requests_current_total", "Current Requests");
         internal readonly MeanCounter FailedRequests = new MeanCounter("failed-requests", "aspnetcore_requests_failed_total", "Failed Requests");
-
-        private readonly Dictionary<string, BaseCounter> _counters;
-
-        public PrometheusAspNetCoreCounterAdapter()
-        {
-            _counters = CounterUtils.GenerateDictionary(this);
-        }
-
-        public void OnCounterEvent(IDictionary<string, object> eventPayload)
-        {
-            if (!eventPayload.TryGetValue("Name", out var counterName))
-            {
-                return;
-            }
-            
-            if (!_counters.TryGetValue((string) counterName, out var counter))
-                return;
-
-            counter.TryReadEventCounterData(eventPayload);
-        }
     }
 }
